requirements.txt
tensorflow-addons
tensorflow-datasets